Move almost all logic of GuiPluginView into EdgeProviderComponent
No functional change
